requirements.txt
langchain>=0.0.267
llama-cpp-python>=0.1.77
transformers>=4.30.2
sentence-transformers>=2.2.2
llama-index>=0.8.0
pyresparser>=1.0.6
requests>=2.31.0
beautifulsoup4>=4.12.2
pandas>=2.0.3
google-api-python-client>=2.95.0
APScheduler>=3.10.1
typer>=0.9.0
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
